split out header file
