+++ arm9/overlays/24/src/overlay_24_c.c
@@ -2,82 +2,13 @@
 #include "bg_window.h"
 #include "unk_0200CA44.h"
 #include "gf_gfx_loader.h"
-
-typedef struct UnkSubStructOverlay24_1 UnkSubStructOverlay24_1;
-struct UnkSubStructOverlay24_1
-{
-    u32 unk0;
-    u32 lastModifiedX;
-    u32 lastModifiedY;
-    u8 pixelData[78][75];
-    u16 unk16E6;
-    u32 unk16E8;
-};
-
-typedef struct UnkSubStructOverlay24 UnkSubStructOverlay24;
-struct UnkSubStructOverlay24
-{
-    UnkSubStructOverlay24_1* unk0;
-    BgConfig* config;
-    u32 unk8;
-    u8 unkC[0x44];
-    u32 unk50;
-    u8 unk54[0x14];
-    u32* unk68[2];
-    Window *window;
-};
-typedef struct UnkStructOverlay24 UnkStructOverlay24;
-struct UnkStructOverlay24
-{
-    u8 unk0;
-    u8 unk1;
-    u8 unk2;
-    u8 unk3;
-    u32 unk4;
-    UnkSubStructOverlay24_1 unk8;
-    UnkSubStructOverlay24* unk16F4;
-    u32 unk16F8;
-    u32 unk16FC;
-};
-
-
-typedef BOOL (*FnType02254918)(UnkStructOverlay24*);
-
-BOOL ov24_02254854(UnkStructOverlay24**, int, int, int);
-BOOL ov24_0225489C(UnkStructOverlay24*, u32, u32, u32);
-void ov24_022548F4(UnkStructOverlay24*);
-void ov24_02254918(void*, UnkStructOverlay24*);
-void ov24_02254960(int, int, int, UnkStructOverlay24*);
-void ov24_02254990(UnkStructOverlay24*);
-void ov24_02254998(UnkStructOverlay24*, u8);
-BOOL ov24_022549AC(UnkStructOverlay24*);
-BOOL ov24_022549F8(UnkStructOverlay24*);
-BOOL ov24_02254A70(UnkStructOverlay24*);
-BOOL ov24_02254AD4(UnkStructOverlay24*);
-void ov24_02254B20(UnkStructOverlay24*, u32, u32, u32, u32);
-BOOL ov24_02254C64(UnkStructOverlay24*);
-BOOL ov24_02254CA0(UnkSubStructOverlay24**, UnkSubStructOverlay24_1*);
-BOOL ov24_02254D00(UnkSubStructOverlay24*);
-void ov24_02254D48(UnkSubStructOverlay24*);
-void ov24_02254D8C(UnkSubStructOverlay24*, u32);
-BOOL ov24_02254DB0(UnkSubStructOverlay24*, u8);
-BOOL ov24_02254DBC(UnkSubStructOverlay24*);
-void ov24_02254DC8(void*);
-void ov24_02254DDC(void*, void*);
-void ov24_02254EE0(u32, void*);
-void ov24_02254F28(int, void*);
-void ov24_02254F40(u32 arg0, void* arg1);
-void ov24_02255038(u32 arg0, void* arg1);
-void ov24_02255050(u32 arg0, void* arg1);
-void ov24_02255078(UnkSubStructOverlay24* arg0);
-void ov24_022550D4(UnkSubStructOverlay24* arg0);
+#include "overlay_24.h"
 
 extern void Poketch_InitApp(void *func1, void *func2);
 extern u32 ov20_02254130(void *, u32, void *, void*, u32);
 extern void ov20_02254198(u32);
 extern void ov20_02252C14(u32, u32);
 extern void ov20_022529A0(u32);
-extern void ov24_0225499(UnkStructOverlay24*);
 extern void ov20_0225298C(u32);
 extern BOOL ov20_02252C08(u32);
 extern BOOL TouchScreen_GetTapState(u32*, u32*);
@@ -92,20 +23,22 @@ extern void ov20_02253F28(u32*, u32);
 extern void Poketch_PlaySoundEffect(u32);
 extern void* ov20_022538A0(void*);
 extern void ov20_02253888(u32*, void*);
-extern BOOL (ov20_022537B8)(void*);
-extern BOOL (ov20_02253794)(void*, u8);
-extern void ov20_022537E0(u32*, u32, void*, void*, void*, u32, u32);
+extern BOOL ov20_022537B8(void*);
+extern BOOL ov20_02253794(void*, u8);
+extern void ov20_022537E0(const u32*, u32, void*, void*, void*, u32, u32);
 extern BOOL ov20_02253FBC(void*, NarcId, u32, u32, HeapID);
 extern u32* ov20_02253E74(u32, void*, void*);
 extern void ov20_02254014(void*);
 extern void ov20_02253F14(u32, void*);
 
-extern FnType02254918 ov24_02255100[3];
-extern u32 ov24_0225514C[2][4];
-extern u32 ov24_0225516C[];
+// TODO: rodata
+extern const FnType02254918 ov24_02255100[3];
+extern const u32 ov24_0225514C[2][4];
+extern const u32 ov24_0225516C[];
+extern const u8 ov24_022550F8[];
 extern const WindowTemplate ov24_0225510C;
-extern BgTemplate ov24_02255114;
-extern BgTemplate ov24_02255130;
+extern const BgTemplate ov24_02255114;
+extern const BgTemplate ov24_02255130;
 
 static void ov24_02254840(void)
 {
